requirements.txt
altgraph==0.16.1
auto-py-to-exe==2.6.5
autopep8==1.4.4
bottle==0.12.17
bottle-websocket==0.2.9
cffi==1.12.3
Click==7.0
cmake==3.17.2
dlib==19.18.0
Eel==0.10.4
face-recognition==1.2.3
face-recognition-models==0.3.0
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
numpy==1.17.2
opencv-python==4.1.1.26
pefile==2019.4.18
Pillow==6.2.0
py2exe==0.9.2.2
pycodestyle==2.5.0
pycparser==2.19
pyinstaller==4.0.dev0
pywin32-ctypes==0.2.0
scipy==1.3.1
uuid==1.30
whichcraft==0.6.1